requirements.txt
boto3>=1.34.0
streamlit>=1.31.0
Pillow>=10.0.0
pdf2image>=1.16.3
python-dotenv>=1.0.0
